refactor(replication_handler): take the data-instance handlers
The data-instance management server handlers live beside the replication handler
they drive, in the namespace they already declared. They receive the
coordinator's RPCs on a data instance and act on the local replication handler,
so they belong to that library; they sat under coordination because they parse
coordinator message types.

mg-coordination no longer links mg-replication_handler, which reduces the cyclic
group of four libraries to the dbms and query pair. Two dependencies that had
been reaching their headers through that link now say so themselves:
coordination's own use of metrics, and three unit tests that include replication
handler headers directly.
