+++ src/replication_handler/CMakeLists.txt
@@ -9,12 +9,14 @@ target_sources(mg-replication_handler
         include/replication_handler/system_replication.hpp
         include/replication_handler/system_rpc.hpp
         include/replication_handler/auth_replication_handlers.hpp
+        include/replication_handler/data_instance_management_server_handlers.hpp
 
         PRIVATE
         system_replication.cpp
         replication_handler.cpp
         system_rpc.cpp
         auth_replication_handlers.cpp
+        data_instance_management_server_handlers.cpp
 )
 
 target_link_libraries(mg-replication_handler
